# docs/api/api.yaml
...
openapi: 3.0.3
info:
  title: Shelfinity Library Management System API
  description: >
    RESTful API for managing library books, users, borrow/return requests,
    reservations, overdue tracking, email configuration, and reports.
    Authentication is delegated entirely to Keycloak (OIDC) — this API validates
    bearer tokens, it does not issue them or store credentials.
  version: 1.0.0
  contact:
    name: Shadow-Codex
servers:
  - url: http://localhost:9080/api
    description: Local development server
...
paths:
...
  /reservations:
    post:
      tags: [Reservations]
      summary: Reserve a currently unavailable book
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReservationRequest' }
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReservationResponse' }
        '400':
          description: Book is currently available, or already reserved by this user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Reservations]
      summary: List all reservations (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Reservations retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReservationResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /reservations/my:
    get:
      tags: [Reservations]
      summary: List the current user's reservations
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Reservations retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReservationResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reservations/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidId'
    delete:
      tags: [Reservations]
      summary: Cancel a reservation (owner or admin)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Reservation cancelled }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reservations/{id}/fulfill:
    parameters:
      - $ref: '#/components/parameters/UuidId'
    post:
      tags: [Reservations]
      summary: Mark a reservation as fulfilled (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Reservation fulfilled }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- Overdue
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued OIDC access token.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted (e.g. admin-only endpoint)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
...
    ReservationStatus:
      type: string
      enum: [ACTIVE, NOTIFIED, FULFILLED, CANCELLED, EXPIRED]
...
    ReservationResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookId: { type: string, format: uuid }
        bookTitle: { type: string }
        userKeycloakId: { type: string }
        status: { $ref: '#/components/schemas/ReservationStatus' }
        notes: { type: string, nullable: true }
        notifiedAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    CreateReservationRequest:
      type: object
      required: [bookId]
      properties:
        bookId: { type: string, format: uuid }
        notes: { type: string }
...